requirements.txt
selenium
pytest
pytest-rerunfailures
allure-pytest
webdriver-manager
